requirements.txt
Flask==2.2.2
Flask_SocketIO==5.3.1
matplotlib==3.8.1
numpy==1.23.5
Pillow==10.1.0
tensorflow==2.12.0
tqdm==4.66.1
translate==3.6.1
